Added types for actions. Split reducers into separate files
