+++ client/src/actions/index.js
@@ -1,9 +1,10 @@
 import axios from 'axios'
+import { GET_BILLS, CREATE_BILL, ITEMS_LOADING, DELETE_BILL, FILTER_BILLS } from './types'
 
 export const deleteBill = id => dispatch => {
   axios.delete(`/api/bills/${id}`).then(res =>
     dispatch({
-      type: 'DELETE_BILL',
+      type: DELETE_BILL,
       payload: id
     })
   )
@@ -12,24 +13,16 @@ export const deleteBill = id => dispatch => {
 export const filterBills = weekFilter => dispatch => {
   axios.get(`/api/bills/${weekFilter}`).then(res =>
     dispatch({
-      type: 'FILTER_BILLS',
+      type: FILTER_BILLS,
       payload: res.data
     })
   )
 }
 
-export function selectBill(bill) {
-  return { type: 'SELECT_BILL', bill }
-}
-
-export function editBill(bill) {
-  return { type: 'EDIT_BILL', bill }
-}
-
 export const createBill = bill => dispatch => {
   axios.post('/api/bills', bill).then(res =>
     dispatch({
-      type: 'CREATE_BILL',
+      type: CREATE_BILL,
       payload: res.data
     })
   )
@@ -39,7 +32,7 @@ export const getBills = () => dispatch => {
   dispatch(setItemsLoading())
   axios.get('/api/bills').then(res => 
     dispatch({
-      type: 'GET_BILLS',
+      type: GET_BILLS,
       payload: res.data
     })
   )
@@ -47,6 +40,6 @@ export const getBills = () => dispatch => {
 
 export const setItemsLoading = () => {
   return {
-    type: 'ITEMS_LOADING'
+    type: ITEMS_LOADING
   }
 }
