+++ client/src/actions/types.js
@@ -1,0 +1,5 @@
+export const DELETE_BILL = 'DELETE_BILL'
+export const FILTER_BILLS = 'FILTER_BILLS'
+export const CREATE_BILL = 'CREATE_BILL'
+export const GET_BILLS = 'GET_BILLS'
+export const ITEMS_LOADING = 'ITEMS_LOADING'
